backend/requirements.txt
# FastAPI Backend
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
geoalchemy2==0.15.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.0
httpx==0.28.0

# Task Queue
celery[redis]==5.4.0
redis==5.2.0

# Settings
pydantic-settings==2.7.0

# Utilities
python-dotenv==1.0.1
orjson==3.10.0

# PDF Reports
weasyprint==62.3
